+++ DV1573---UD1448/Renderer/Renderer.h
@@ -66,7 +66,7 @@ private:
 	//Storage Buffer for light indecies
 	unsigned int m_lightIndexSSBO;
 	
-	glm::vec2 workGroups;
+	glm::uvec2 workGroups;
 	std::vector<Pointlight> m_pLights;//The size of the vector is the number of lights
 	
 	void createDepthMap();
